WIP: move tests around
