+++ tests/providers/gcp/gcp_managed_relational_db_test.py
@@ -1,0 +1,52 @@
+# Copyright 2017 PerfKitBenchmarker Authors. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Tests for perfkitbenchmarker.providers.gcp.gcp_managed_relational_db"""
+
+import contextlib
+import mock
+import re
+import unittest
+
+from perfkitbenchmarker import benchmark_spec
+from perfkitbenchmarker import context
+from perfkitbenchmarker import errors
+from perfkitbenchmarker import os_types
+from perfkitbenchmarker import providers
+from perfkitbenchmarker import vm_util
+from perfkitbenchmarker.configs import benchmark_config_spec
+from perfkitbenchmarker.providers.gcp import gcp_managed_relational_db
+from tests import mock_flags
+
+
+_BENCHMARK_NAME = 'name'
+_BENCHMARK_UID = 'benchmark_uid'
+_COMPONENT = 'test_component'
+_FLAGS = None
+
+
+class GcpManagedRelationalDbSpecTestCase(unittest.TestCase):
+  pass
+
+
+class GceManagedRelationalDbFlagsTestCase(unittest.TestCase):
+  pass
+
+
+class GceManagedRelationalDbTestCase(unittest.TestCase):
+  pass
+
+
+if __name__ == '__main__':
+  unittest.main()
